llm/requirements.txt
vllm[serve]
fastapi
uvicorn
psycopg2-binary
huggingface_hub